Remove old Home feature
