+++ Sources/PointFreeRouter/Routes.swift
@@ -33,7 +33,6 @@ public indirect enum SiteRoute: Equatable {
   case gitHubAuth(redirect: String?)
   case gitHubCallback(code: String?, redirect: String?)
   case home
-  case homeV2
   case invite(Invite)
   case live(Live)
   case login(redirect: String?)
@@ -438,9 +437,6 @@ struct SiteRouter: ParserPrinter {
   var body: some Router<SiteRoute> {
     OneOf {
       Route(.case(SiteRoute.home))
-      Route(.case(SiteRoute.homeV2)) {
-        Path { "home" }
-      }
 
       Route(.case(SiteRoute.about)) {
         Path { "about" }
