+++ Sources/Views/Blog/BlogPostShow.swift
@@ -13,11 +13,6 @@ public func blogPostShowView(post: BlogPost) -> Node {
   @Dependency(\.date.now) var now
   @Dependency(\.subscriberState) var subscriberState
 
-  let showHolidaySpecialCallout =
-    holidayDiscount2019Interval.contains(now.timeIntervalSince1970)
-    && subscriberState.isNonSubscriber
-    && post.id != 36
-
   return [
     .gridRow(
       attributes: [.class([Class.padding([.mobile: [.leftRight: 3], .desktop: [.leftRight: 4]])])],
@@ -25,13 +20,12 @@ public func blogPostShowView(post: BlogPost) -> Node {
         .gridColumn(
           sizes: [.mobile: 12, .desktop: 9],
           attributes: [.style(margin(leftRight: .auto))],
-          showHolidaySpecialCallout ? holidaySpecialCallout : [],
           .div(
             attributes: [
               .class([Class.padding([.mobile: [.topBottom: 3], .desktop: [.topBottom: 4]])])
             ],
             blogPostContentView(post),
-            subscriberCalloutView
+            postSubscriberCalloutView
           )
         )
       ]
@@ -78,7 +72,7 @@ public func blogPostContentView(_ post: BlogPost) -> Node {
   ]
 }
 
-var subscriberCalloutView: Node {
+private var postSubscriberCalloutView: Node {
   @Dependency(\.subscriberState) var subscriberState
   guard !subscriberState.isActive else { return [] }
 
@@ -119,15 +113,6 @@ var subscriberCalloutView: Node {
     ),
   ]
 }
-
-private let holidaySpecialCallout: Node = .div(
-  attributes: [
-    .class([
-      Class.margin([.mobile: [.top: 4], .desktop: [.leftRight: 4]])
-    ])
-  ],
-  holidaySpecialContent
-)
 
 let episodeDateFormatter: DateFormatter = {
   let df = DateFormatter()
